+++ app/src/main/java/org/simple/clinic/ClinicApp.kt
@@ -4,9 +4,19 @@ import android.annotation.SuppressLint
 import android.app.Application
 import androidx.camera.camera2.Camera2Config
 import androidx.camera.core.CameraXConfig
+import com.datadog.android.Datadog
+import com.datadog.android.core.configuration.Configuration
+import com.datadog.android.core.configuration.Credentials
+import com.datadog.android.privacy.TrackingConsent
+import com.datadog.android.rum.GlobalRum
+import com.datadog.android.rum.RumMonitor
+import com.datadog.android.rum.tracking.FragmentViewTrackingStrategy
+import com.datadog.android.tracing.AndroidTracer
+import io.opentracing.util.GlobalTracer
 import io.reactivex.exceptions.UndeliverableException
 import io.reactivex.plugins.RxJavaPlugins
 import org.simple.clinic.activity.CloseActivitiesWhenUserIsUnauthorized
+import org.simple.clinic.analytics.ResolveScreenNamesForDatadog
 import org.simple.clinic.analytics.UpdateAnalyticsUserId
 import org.simple.clinic.crash.CrashBreadcrumbsTimberTree
 import org.simple.clinic.crash.SentryCrashReporterSink
@@ -14,8 +24,11 @@ import org.simple.clinic.di.AppComponent
 import org.simple.clinic.platform.analytics.Analytics
 import org.simple.clinic.platform.analytics.AnalyticsReporter
 import org.simple.clinic.platform.crash.CrashReporter
+import org.simple.clinic.remoteconfig.ConfigReader
 import org.simple.clinic.storage.monitoring.AnalyticsSqlPerformanceReportingSink
+import org.simple.clinic.storage.monitoring.DatadogSqlPerformanceReportingSink
 import org.simple.clinic.storage.monitoring.SqlPerformanceReporter
+import org.simple.clinic.util.clamp
 import timber.log.Timber
 import java.io.IOException
 import java.net.SocketException
@@ -39,6 +52,9 @@ abstract class ClinicApp : Application(), CameraXConfig.Provider {
   @Inject
   lateinit var sentryCrashReporterSink: SentryCrashReporterSink
 
+  @Inject
+  lateinit var remoteConfig: ConfigReader
+
   protected open val analyticsReporters = emptyList<AnalyticsReporter>()
 
   protected open val crashReporterSinks = emptyList<CrashReporter.Sink>()
@@ -51,6 +67,9 @@ abstract class ClinicApp : Application(), CameraXConfig.Provider {
     appComponent.inject(this)
 
     crashReporterSinks.forEach(CrashReporter::addSink)
+
+    setupApplicationPerformanceMonitoring()
+
     Timber.plant(CrashBreadcrumbsTimberTree())
     RxJavaPlugins.setErrorHandler { error ->
       if (!error.canBeIgnoredSafely()) {
@@ -63,11 +82,45 @@ abstract class ClinicApp : Application(), CameraXConfig.Provider {
       Analytics.addReporter(reporter)
     }
     SqlPerformanceReporter.addSink(analyticsSqlPerformanceReportingSink)
+    SqlPerformanceReporter.addSink(DatadogSqlPerformanceReportingSink())
 
     updateAnalyticsUserId.listen()
 
     registerActivityLifecycleCallbacks(closeActivitiesWhenUserIsUnauthorized)
     closeActivitiesWhenUserIsUnauthorized.listen()
+  }
+
+  private fun setupApplicationPerformanceMonitoring() {
+    val samplingRate = remoteConfig
+        .double("datadog_sample_rate", 0.0)
+        .toFloat()
+        .clamp(0F, 100F)
+
+    val datadogConfig = Configuration
+        .Builder(
+            logsEnabled = false,
+            tracesEnabled = true,
+            crashReportsEnabled = false,
+            rumEnabled = true
+        )
+        .trackBackgroundRumEvents(true)
+        .trackLongTasks(5000)
+        .useViewTrackingStrategy(FragmentViewTrackingStrategy(
+            trackArguments = false,
+            supportFragmentComponentPredicate = ResolveScreenNamesForDatadog()
+        ))
+        .sampleRumSessions(samplingRate = samplingRate)
+        .build()
+    val credentials = Credentials(
+        clientToken = BuildConfig.DATADOG_CLIENT_TOKEN,
+        envName = BuildConfig.DATADOG_ENVIRONMENT,
+        variant = BuildConfig.FLAVOR,
+        rumApplicationId = BuildConfig.DATADOG_APPLICATION_ID,
+        serviceName = BuildConfig.DATADOG_SERVICE_NAME
+    )
+    Datadog.initialize(this, credentials, datadogConfig, TrackingConsent.GRANTED)
+    GlobalRum.registerIfAbsent(RumMonitor.Builder().build())
+    GlobalTracer.registerIfAbsent(AndroidTracer.Builder().build())
   }
 
   override fun getCameraXConfig(): CameraXConfig {